v3 - Show Benefits by location (#609)
* show Benefits by location

* cleanup in benefit schema and moved filtering of benefits to compensatins

* Update studio/schemas/objects/compensations/benefitsByLocation.ts


* created reference interface

---------
