+++ src/compensations/compensations.module.css
@@ -4,16 +4,3 @@
   padding: 10rem 5rem;
   gap: 5rem;
 }
-
-.benefits {
-  display: flex;
-  flex-direction: column;
-  gap: 5rem;
-  max-width: var(--max-content-width-medium);
-}
-
-.benefitWrapper {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
